Added notes to dropdown and decoupled modal from pill button/icon
