+++ frontend/src/views/admin/matching/applicant-view/grid/modals.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Application } from "../../../../../api/defs/types";
-import { MatchableAssignment } from "../../types";
-import { Modal, Button } from "react-bootstrap";
+import { ApplicantSummary, MatchableAssignment } from "../../types";
+import { Form, Modal, Button } from "react-bootstrap";
 import { ApplicationDetails } from "../../../applications/application-details";
-import { upsertMatch } from "../../actions";
+import { upsertMatch, upsertNote } from "../../actions";
 import { useThunkDispatch } from "../../../../../libs/thunk-dispatch";
 
 /**
@@ -96,6 +96,74 @@ export function AdjustHourModal({
                                 stagedHoursAssigned: Number(hoursAssigned),
                             })
                         );
+                        setShow(false);
+                    }}
+                    variant="outline-primary"
+                >
+                    Save
+                </Button>
+            </Modal.Footer>
+        </Modal>
+    );
+}
+
+/**
+ * A modal window allowing users to view and edit notes for an applicant.
+ */
+export function ApplicantNoteModal({
+    applicantSummary,
+    show,
+    setShow,
+}: {
+    applicantSummary: ApplicantSummary;
+    show: boolean;
+    setShow: (arg0: boolean) => void;
+}) {
+    const dispatch = useThunkDispatch();
+    const [noteTemp, setNoteTemp] = React.useState(applicantSummary.note || "");
+
+    const updateApplicantNote = React.useCallback(
+        (note: string) => {
+            dispatch(
+                upsertNote({
+                    utorid: applicantSummary.applicant.utorid,
+                    note: noteTemp,
+                })
+            );
+        },
+        [dispatch, noteTemp, applicantSummary]
+    );
+
+    return (
+        <Modal show={show} onHide={() => setShow(false)}>
+            <Modal.Header closeButton>
+                <Modal.Title>
+                    Notes ({applicantSummary.applicant.first_name}{" "}
+                    {applicantSummary.applicant.last_name})
+                </Modal.Title>
+            </Modal.Header>
+            <Modal.Body>
+                <Form>
+                    <Form.Group className="mb-3">
+                        <Form.Control
+                            as="textarea"
+                            rows={3}
+                            defaultValue={applicantSummary.note || ""}
+                            onChange={(e) => setNoteTemp(e.target.value)}
+                        />
+                    </Form.Group>
+                </Form>
+            </Modal.Body>
+            <Modal.Footer>
+                <Button
+                    onClick={() => setShow(false)}
+                    variant="outline-secondary"
+                >
+                    Cancel
+                </Button>
+                <Button
+                    onClick={() => {
+                        updateApplicantNote(noteTemp);
                         setShow(false);
                     }}
                     variant="outline-primary"
